--- src/main/java/kitchenpos/repository/JpaMenuProductRepository.java
@@ -1,7 +1,0 @@
-package kitchenpos.repository;
-
-import kitchenpos.domain.MenuProduct;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaMenuProductRepository extends JpaRepository<MenuProduct, Long> {
-}
