--- src/main/java/kitchenpos/ui/ProductRestController.java
@@ -1,38 +1,0 @@
-package kitchenpos.ui;
-
-import java.net.URI;
-import java.util.List;
-import kitchenpos.application.ProductService;
-import kitchenpos.domain.Product;
-import kitchenpos.dto.request.ProductCreateRequest;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ProductRestController {
-
-    private final ProductService productService;
-
-    public ProductRestController(final ProductService productService) {
-        this.productService = productService;
-    }
-
-    @PostMapping("/api/products")
-    public ResponseEntity<Product> create(@RequestBody final ProductCreateRequest request) {
-        final Product created = productService.create(request);
-        final URI uri = URI.create("/api/products/" + created.id());
-        return ResponseEntity.created(uri)
-                .body(created)
-                ;
-    }
-
-    @GetMapping("/api/products")
-    public ResponseEntity<List<Product>> list() {
-        return ResponseEntity.ok()
-                .body(productService.list())
-                ;
-    }
-}
